remove browser pool
