+++ src/notte/agents/main.py
@@ -3,7 +3,6 @@ from collections.abc import Callable
 
 from notte.agents.falco.agent import FalcoAgent, FalcoAgentConfig
 from notte.agents.falco.types import StepAgentOutput
-from notte.browser.pool.base import BaseBrowserPool
 from notte.browser.window import BrowserWindow
 from notte.common.agent.base import BaseAgent
 from notte.common.agent.types import AgentResponse
@@ -37,7 +36,6 @@ class Agent:
 
     def create_agent(
         self,
-        pool: BaseBrowserPool | None = None,
         step_callback: Callable[[str, StepAgentOutput], None] | None = None,
         window: BrowserWindow | None = None,
     ) -> BaseAgent:
@@ -45,7 +43,6 @@ class Agent:
             config=self.config,
             vault=self.vault,
             window=window,
-            pool=pool,
             step_callback=step_callback,
         )
         if self.notifier:
