+++ src/notte_integrations/remote_sessions/browserbase_pool.py
@@ -3,8 +3,8 @@ import os
 from loguru import logger
 from typing_extensions import override
 
-from notte.browser.pool.base import BrowserResourceOptions, BrowserWithContexts
-from notte.browser.pool.cdp_pool import BrowserEnum, CDPBrowserPool, CDPSession
+from notte_pools.base import BrowserResourceOptions, BrowserWithContexts
+from notte_pools.cdp_pool import BrowserEnum, CDPBrowserPool, CDPSession
 
 try:
     from browserbase import Browserbase
